audio-transcription/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.24.0.post1
google-cloud-firestore==2.13.1
google-cloud-tasks==2.18.0
google-cloud-logging==3.12.1
